+++ src/aux-server/aux-web/aux-player/scene/PlayerGame.ts
@@ -36,6 +36,9 @@ export class PlayerGame extends Game {
     inventoryCameraRig: CameraRig = null;
     inventoryViewport: Viewport = null;
 
+    startZoom: number;
+    startAspect: number;
+
     private inventoryScene: Scene;
 
     inventoryHeightOverride: number = null;
@@ -43,6 +46,8 @@ export class PlayerGame extends Game {
     private slider: Element;
     private sliderVis: Element;
     private sliderPressed: boolean = false;
+
+    setupDelay: boolean = false;
 
     constructor(gameView: PlayerGameView) {
         super(gameView);
@@ -347,6 +352,8 @@ export class PlayerGame extends Game {
         // Inventory direction light.
         const invDirectional = baseAuxDirectionalLight();
         this.inventoryScene.add(invDirectional);
+
+        this.setupDelay = true;
     }
 
     onWindowResize(width: number, height: number) {
@@ -408,6 +415,13 @@ export class PlayerGame extends Game {
 
     async mouseDownSlider() {
         this.sliderPressed = true;
+
+        if (this.inventoryCameraRig.mainCamera instanceof OrthographicCamera) {
+            this.startAspect =
+                this.inventoryCameraRig.viewport.width /
+                this.inventoryCameraRig.viewport.height;
+            this.startZoom = this.inventoryCameraRig.mainCamera.zoom;
+        }
     }
 
     async mouseUpSlider() {
@@ -422,6 +436,11 @@ export class PlayerGame extends Game {
 
     frameUpdate() {
         super.frameUpdate();
+
+        if (this.setupDelay) {
+            this.onCenterCamera(this.inventoryCameraRig);
+            this.setupDelay = false;
+        }
 
         if (!this.sliderPressed) return false;
 
@@ -466,9 +485,17 @@ export class PlayerGame extends Game {
         if (cameraRig.mainCamera instanceof OrthographicCamera) {
             const aspect = cameraRig.viewport.width / cameraRig.viewport.height;
 
-            // found that 50 is the preset zoom of the rig.maincamera.zoom so I am using this as the base zoom
-            const newZoom = 50 - (49 - aspect * 7);
-            cameraRig.mainCamera.zoom = newZoom;
+            if (this.startAspect != null) {
+                let zoomC = this.startZoom / this.startAspect;
+                const newZoom =
+                    this.startZoom - (this.startZoom - aspect * zoomC);
+                cameraRig.mainCamera.zoom = newZoom;
+            } else {
+                let initNum = 80;
+                // found that 50 is the preset zoom of the rig.maincamera.zoom so I am using this as the base zoom
+                const newZoom = initNum - (initNum - aspect * (initNum / 7));
+                cameraRig.mainCamera.zoom = newZoom;
+            }
         }
     }
 }