+++ src/aux-server/aux-web/shared/scene/Game.ts
@@ -322,6 +322,14 @@ export abstract class Game implements AuxFile3DFinder {
 
     onCenterCamera(cameraRig: CameraRig): void {
         if (!cameraRig) return;
+
+        let controls = this.interaction.cameraRigControllers.find(
+            c => c.rig.name === cameraRig.name
+        );
+
+        controls.controls.resetRot = true;
+        controls.controls.update();
+
         this.tweenCameraToPosition(cameraRig, new Vector3(0, 0, 0));
     }